requirements.txt
numpy==1.24.3
pandas==2.0.2
python-dotenv==1.0.0
python_docx==1.1.0
PyYAML==6.0.1
selenium==4.16.0
undetected_chromedriver==3.5.4
